+++ projects/ng-ga4/src/public-api.ts
@@ -1,4 +1,5 @@
 export { NgGa4Config, NgGa4CookieOptions, NG_GA4_CONFIG } from './lib/ng-ga4.config';
+export { NgGa4Consent, NgGa4ConsentState } from './lib/consent';
 export { NgGa4Service } from './lib/ng-ga4.service';
 export { NgGa4Module } from './lib/ng-ga4.module';
 export { provideNgGa4 } from './lib/ng-ga4.providers';